Now tested with a home-made Blender exporter
Which is a bit more useful than a Unity exporter.
If the mesh you need is already on Unity, chances
are that you can upload it to your world very quickly.

If it's on Blender, the process is longer and a
bit more painful (though, you get everything and not
just a mesh with one single material).

Signed-off-by: Voyage <voyage@miouyouyou.fr>