DMI : Multiple materials support on the way !
The panel is still not ready, there's still a few
bugs here and there, but for what I've tested, it works
fine !

You can now setup as much textures as the shader
supoprts it !

And you can also manage multiple materials.

Supporting multiple materials quickly became a pain
since, for some reasons, VRChat decided to blacklist
the access to SubMeshDescriptor, which is normally
REQUIRED to define submeshes, which are in turn essential
to support multiple materials.

Basically, each material apply to each submesh.

That said, I was able to abuse CombineInstances and
Mesh.Optimize() to actually generate a mesh with all
the materials and all the submeshes required.
It's a horrible hack but kind of works, and isn't THAT
heavy (but isn't THAT light either).

Anyway, what's remaining is to test various quirks
with the UI.
This will require some VRChat API mocks at some point,
since testing using the Unity Editor alone is a PAIN !

Signed-off-by: Voyage <voyage@miouyouyou.fr>