Materials are now instantiated correctly
And I started to figure out how to setup the remaining textures
slots correctly.
I just need to preload them in advance, since enabling shader
keywords is not trivial.

Shader keywords are not trivial since their names are not
related to the texture units they control.
For example, to be able to set the _BumpMap slot correctly,
you need to enable the kewyord _NORMALMAP.

So, in order to avoid having to figure out all the potentials
keywords combinations, I'll just preload all the slots with the
default textures defined in the properties
(white, gray, black, normal). That way, we should be good to go.

Signed-off-by: Voyage <voyage@miouyouyou.fr>